requirements.txt
django>=4.0
pillow
django-ckeditor
django-modeltranslation
django-crispy-forms
crispy-bootstrap5
